Merges develop branch into master by switching to the git of Clap and refactoring function calls
