+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "leftwm-theme"
 version = "0.0.1"
-authors = ["Mautamu <mautam@usa.com>"]
+authors = ["Lex Childs <lexchilds@gmail.com>", "Hitesh Paul <git@hiteshpaul.com>", "Mautamu <mautam@usa.com>"]
 description = "A theme mangager for LeftWM"
 edition = "2018"
 keywords = ["leftwm", "wm", "theme"]
@@ -12,7 +12,8 @@ readme = "README.md"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-clap = "2.33.0"
+clap = { git = "https://github.com/clap-rs/clap/", features=["derive"] }
+#clap = {version = "2.33.0", features = ["derive"]}
 reqwest = {version = "0.11.0", features= ["blocking", "json"]}
 dirs-next = "2.0.0"
 toml = "0.5.5"
