--- src/functions/mod.rs
@@ -1,15 +1,0 @@
-mod apply;
-mod install;
-mod list;
-mod new;
-mod update;
-mod uninstall;
-mod upgrade;
-
-pub use apply::apply;
-pub use install::install;
-pub use list::list;
-pub use new::new;
-pub use update::update;
-pub use uninstall::uninstall;
-pub use upgrade::upgrade;
