+++ src/operations/mod.rs
@@ -1,0 +1,17 @@
+mod apply;
+mod install;
+mod list;
+mod new;
+mod status;
+mod uninstall;
+mod update;
+mod upgrade;
+
+pub use apply::Apply;
+pub use install::Install;
+pub use list::List;
+pub use new::New;
+pub use status::Status;
+pub use uninstall::Uninstall;
+pub use update::Update;
+pub use upgrade::Upgrade;
